requirements.txt
pytube
whisper
transformers
torch
python-docx
fpdf